@@ -5,7 +5,6 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve --proxy-config proxy.config.json",
-    "start:hmr": "ng serve --proxy-config proxy.config.json --hmr -e=hmr",
     "start:fr": "ng serve --aot --i18nFile=src/locale/messages.fr.xlf --i18nFormat=xlf --locale=fr --proxy-config proxy.config.json",
     "build": "ng build --aot --prod --build-optimizer",
     "analyze": "ng build --stats-json && webpack-bundle-analyzer dist/stats.json",
@@ -16,56 +15,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.4",
-    "@angular/common": "^5.2.4",
-    "@angular/compiler": "^5.2.4",
-    "@angular/core": "^5.2.4",
-    "@angular/forms": "^5.2.4",
-    "@angular/http": "^5.2.4",
-    "@angular/language-service": "^5.2.4",
-    "@angular/platform-browser": "^5.2.4",
-    "@angular/platform-browser-dynamic": "^5.2.4",
-    "@angular/router": "^5.2.4",
-    "@angular/service-worker": "^5.2.4",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "@fortawesome/fontawesome": "^1.1.3",
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
-    "@ng-select/ng-select": "^0.28.1",
+    "@ng-select/ng-select": "^1.4.2",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^5.0.0",
+    "@ngrx/effects": "^5.2.0",
     "@ngrx/router-store": "^5.0.0",
-    "@ngrx/store": "^5.0.0",
-    "@ngrx/store-devtools": "^5.0.0",
-    "@types/uuid": "^3.4.3",
-    "core-js": "^2.4.1",
+    "@ngrx/store": "^5.2.0",
+    "@ngrx/store-devtools": "^5.2.0",
+    "apollo-angular": "^1.0.1",
+    "apollo-angular-cache-ngrx": "^1.0.0-beta.0",
+    "apollo-angular-link-http": "^1.0.3",
+    "apollo-cache-inmemory": "^1.2.0",
+    "apollo-client": "^2.3.0",
+    "apollo-link-ws": "^1.0.8",
+    "core-js": "^2.5.4",
+    "graphql": "^0.13.2",
+    "graphql-tag": "^2.9.2",
     "ngrx-store-freeze": "^0.2.0",
     "ngrx-store-logger": "^0.2.0",
-    "rxjs": "^5.5.7",
+    "rxjs": "^6.0.0",
+    "rxjs-compat": "^6.1.0",
+    "subscriptions-transport-ws": "^0.9.9",
     "uuid": "^3.2.1",
-    "webpack": "^3.11.0",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/core": "^0.4.2",
-    "@angular-devkit/schematics": "^0.4.2",
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.4",
-    "@angularclass/hmr": "^2.1.3",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "typescript": "~2.7.2",
+    "@angular/cli": "~6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "^2.4.2",
-    "webpack-bundle-analyzer": "^2.9.2"
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
+    "webpack-bundle-analyzer": "^2.9.2",
+    "@types/uuid": "^3.4.3"
   }
 }
