@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable ,  forkJoin } from 'rxjs';
 import { map, mergeMap, switchMap, withLatestFrom } from 'rxjs/operators';
 import { EntityService } from '~entity/store/entity.service';
 import { ERM } from '~entity/store/entity.model';
@@ -9,7 +9,6 @@ import { selectUserTeamId } from '../user';
 
 import { ProjectHttpService } from './project-http.service';
 import { fromProject } from './project.bundle';
-import { forkJoin } from 'rxjs/observable/forkJoin';
 
 
 const actionType = fromProject.ActionTypes;
