@@ -3,7 +3,7 @@ import { FormGroup, FormBuilder } from '@angular/forms';
 import { Validators } from '@angular/forms';
 import { Store } from '@ngrx/store';
 import * as fromAuth from '~app/features/auth/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { selectResetPwPagePending, selectResetPwPageError } from '~app/features/auth/store';
 
 @Component({
