@@ -3,12 +3,11 @@ import { Injectable } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
-import { debounceTime, map, mergeMap, startWith, switchMap, tap, filter } from 'rxjs/operators';
-import { catchError } from 'rxjs/operators/catchError';
+import { debounceTime, map, mergeMap, startWith, switchMap, tap, filter ,  catchError } from 'rxjs/operators';
 import { UserActions } from '~user';
 import { AuthHttpService, TokenService } from '~auth/services';
 import { TypedAction } from '~app/app-root/utils';
-import { of } from 'rxjs/observable/of';
+import { of } from 'rxjs';
 import { LocalStorageService } from '~app/shared/local-storage';
 import { AuthActionType, AuthActions } from './authentication.action';
 import { PreloaderActions } from '~app/shared/preloader/preloader.action';
