@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { selectEntityArray } from '~entity';
 import { Entity, EntityRepresentation, ERM } from '~entity';
 
