@@ -2,7 +2,7 @@ import { Component, OnInit, ChangeDetectionStrategy, Input } from '@angular/core
 import { Product } from '~feature/products/store/product/product.model';
 import { Store } from '@ngrx/store';
 import { LatestProductActions, selectLatestProductsArray } from '~app/features/supplier/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Component({
 	selector: 'supplier-latest-products-app',
