@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Entity, EntityTarget, FormDescriptor, FormGroupDescriptor, selectEntityById } from '~entity';
 import { AutoUnsub } from '~utils';
 
