@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { AppFile, fromFile } from '~entity';
 import { AutoUnsub } from '~utils';
 import { UserService } from '~app/features/user';
