@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { selectUser, User } from '~user';
 import { AuthActions } from '~app/features/auth';
 import { Team, fromTeam } from '~app/entity';
