@@ -2,7 +2,6 @@ import { AfterViewInit, Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 
 import { Log } from '~utils';
-import { HmrService } from '~app/shared/hmr/hmr.service';
 import { AuthActions } from '~app/features/auth';
 
 @Component({
@@ -11,13 +10,11 @@ import { AuthActions } from '~app/features/auth';
 	styleUrls: ['./app.component.scss'],
 })
 export class AppComponent implements OnInit {
-	constructor(private store: Store<any>, private hmrService: HmrService) { }
+	constructor(private store: Store<any>) { }
 
 	ngOnInit(): void {
 		Log.info('App init');
 		// hmr service or hot module reloading
-		if (!this.hmrService.isStoreLoaded()) {
-			this.store.dispatch(AuthActions.checkAuthenticated());
-		}
+		this.store.dispatch(AuthActions.checkAuthenticated());
 	}
 }
