@@ -1,44 +1,39 @@
 import { Category } from '../category';
 import { Tag } from '../tag';
-import { uuid } from '~utils';
+import { AppImage } from '~app/entity';
 
-export class Supplier {
-	favorite: boolean;
-	status: string;
-	supplierType: string;
-	categories: Array<Category>;
-	counters: any;
-	countryCode: string;
-	createdByUserId: string;
-	creationDate: number;
-	id: string;
-	lastModifiedDate: number;
-	lastUpdatedByUserId: string;
-	name: string;
-	rating: number;
-	tags: Array<Tag>;
-	teamId: string;
-	email: string;
-	website: string;
-	phone: string;
-	description: string;
-	pending = true;
-	minimumOrderQuantity: number;
-	leadTime: number;
-	phoneNumber: string;
-	address: string;
-	type: string;
-	incoTerm: string;
-	harbour: string;
-	eventId: string;
-	categoryIds: Array<String>;
-	tagIds: Array<String>;
-	leadTimeValue: number;
+// TODO audit is used everywhere move it where it makes sens
+export interface Audit {
+	createdBy?: any;
+	creationDate?: any;
+}
 
-	constructor(name: string, userId: string) {
-		this.name = name;
-		this.createdByUserId = userId;
-		this.creationDate = Date.now();
-		this.id = uuid();
-	}
+export interface Supplier {
+	id?: string;
+	name?: string;
+	fullName?: string;
+	description?: string;
+	images?: AppImage[];
+	logoImage?: AppImage;
+	supplierType?: SupplierType;
+	website?: string;
+	phoneNumber?: string;
+	country?: string;
+	address?: string;
+	officeEmail?: string;
+	officePhone?: string;
+	incoTerm?: string;
+	harbour?: string;
+	generalMOQ?: number;
+	generalLeadTime?: number;
+	tags?: Tag[];
+	categories?: Category[];
+	favorite?: boolean;
+	globalDatabaseId?: string;
+	audit?: Audit;
+	deleted?: boolean;
+}
+
+export enum SupplierType {
+
 }
