@@ -4,12 +4,10 @@ import { map, switchMap, mergeMap, catchError, tap } from 'rxjs/operators';
 import { NotificationType } from '~app/shared/notifications/model/notification.interface';
 import { notificationActions } from '~app/shared/notifications/store/notification.action';
 import { appErrorActions } from '~app/shared/error-handler/app-errors.action';
-import { of } from 'rxjs/observable/of';
+import { of ,  merge ,  from } from 'rxjs';
 import { DialogActions } from '~app/shared/dialog/store/dialog.action';
 import { DialogName } from '~app/shared/dialog';
 import { Router } from '@angular/router';
-import { merge } from 'rxjs/observable/merge';
-import { from } from 'rxjs/observable/from';
 import { NewProductDlgActionType, NewProductDlgActions } from '~app/features/products/store/new-product-dlg/new-product-dlg.actions';
 import { productActions } from '~app/features/products/store/product';
 import { ProductHttpService } from '~app/features/products/store/product/product-http.service';
