@@ -1,8 +1,13 @@
+import { AppImage, Supplier } from '~app/entity';
 
 export interface Contact {
-	id: string;
-	name: string;
-	email: string;
-	phoneNumber: string;
-	supplierId: string;
+	id?: string;
+	name?: string;
+	phoneNumber?: string;
+	email?: string;
+	businessCardImage?: AppImage;
+	jobTitle?: string;
+	supplier?: Supplier;
+	audit?: any;
+	deleted?: boolean;
 }
