@@ -6,13 +6,11 @@ import { NotificationType } from '~app/shared/notifications/model/notification.i
 import { notificationActions } from '~app/shared/notifications/store/notification.action';
 import { appErrorActions } from '~app/shared/error-handler/app-errors.action';
 import { SupplierHttpService } from '~app/entity/store/supplier/supplier-http.service';
-import { of } from 'rxjs/observable/of';
+import { of ,  merge ,  from } from 'rxjs';
 import { supplierActions } from '~app/entity/store/supplier/supplier.action';
 import { DialogActions } from '~app/shared/dialog/store/dialog.action';
 import { DialogName } from '~app/shared/dialog';
 import { Router } from '@angular/router';
-import { merge } from 'rxjs/observable/merge';
-import { from } from 'rxjs/observable/from';
 
 @Injectable()
 export class NewSupplierDlgEffects {
