@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Notification } from '../model/notification.interface';
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 
 @Injectable()
 export class NotificationService {
